разделил статистику на функции
