+++ src/stats-render.js
@@ -1,0 +1,120 @@
+import Chart from 'chart.js';
+import ChartDataLabels from 'chartjs-plugin-datalabels';
+
+export const renderColorsStats = (colorLabels, cardColorsCount) => {
+  const colorsCtx = document.querySelector(`.statistic__colors`);
+
+  // eslint-disable-next-line no-unused-vars
+  const colorsChart = new Chart(colorsCtx, {
+    plugins: [ChartDataLabels],
+    type: `pie`,
+    data: {
+      labels: colorLabels,
+      datasets: [{
+        data: cardColorsCount,
+        backgroundColor: [`#ff3cb9`, `#ffe125`, `#0c5cdd`, `#000000`, `#31b55c`]
+      }]
+    },
+    options: {
+      plugins: {
+        datalabels: {
+          display: false
+        }
+      },
+      tooltips: {
+        callbacks: {
+          label: (tooltipItem, data) => {
+            const allData = data.datasets[tooltipItem.datasetIndex].data;
+            const tooltipData = allData[tooltipItem.index];
+            const total = allData.reduce((acc, it) => acc + parseFloat(it));
+            const tooltipPercentage = Math.round((tooltipData / total) * 100);
+            return `${tooltipData} TASKS — ${tooltipPercentage}%`;
+          }
+        },
+        displayColors: false,
+        backgroundColor: `#ffffff`,
+        bodyFontColor: `#000000`,
+        borderColor: `#000000`,
+        borderWidth: 1,
+        cornerRadius: 0,
+        xPadding: 15,
+        yPadding: 15
+      },
+      title: {
+        display: true,
+        text: `DONE BY: COLORS`,
+        fontSize: 16,
+        fontColor: `#000000`
+      },
+      legend: {
+        position: `left`,
+        labels: {
+          boxWidth: 15,
+          padding: 25,
+          fontStyle: 500,
+          fontColor: `#000000`,
+          fontSize: 13
+        }
+      }
+    }
+  });
+};
+
+export const renderTagsStats = (tags, tagsCountArr, tagsColors) => {
+  const tagsCtx = document.querySelector(`.statistic__tags`);
+
+  // eslint-disable-next-line no-unused-vars
+  const tagsChart = new Chart(tagsCtx, {
+    plugins: [ChartDataLabels],
+    type: `pie`,
+    data: {
+      labels: tags,
+      datasets: [{
+        data: tagsCountArr,
+        backgroundColor: tagsColors
+      }]
+    },
+    options: {
+      plugins: {
+        datalabels: {
+          display: false
+        }
+      },
+      tooltips: {
+        callbacks: {
+          label: (tooltipItem, data) => {
+            const allData = data.datasets[tooltipItem.datasetIndex].data;
+            const tooltipData = allData[tooltipItem.index];
+            const total = allData.reduce((acc, it) => acc + parseFloat(it));
+            const tooltipPercentage = Math.round((tooltipData / total) * 100);
+            return `${tooltipData} TASKS — ${tooltipPercentage}%`;
+          }
+        },
+        displayColors: false,
+        backgroundColor: `#ffffff`,
+        bodyFontColor: `#000000`,
+        borderColor: `#000000`,
+        borderWidth: 1,
+        cornerRadius: 0,
+        xPadding: 15,
+        yPadding: 15
+      },
+      title: {
+        display: true,
+        text: `DONE BY: TAGS`,
+        fontSize: 16,
+        fontColor: `#000000`
+      },
+      legend: {
+        position: `left`,
+        labels: {
+          boxWidth: 15,
+          padding: 25,
+          fontStyle: 500,
+          fontColor: `#000000`,
+          fontSize: 13
+        }
+      }
+    }
+  });
+}
